+++ CHANGELOG.rst
@@ -92,7 +92,8 @@ Fixed
     - Timeline event status not updated in ``SheetDeleteVieW`` with iRODS collections enabled (#1798)
     - Assay plugin ``update_row()`` setting links for empty file names (#2017)
     - Sporadic test failure in ``TestIrodsAccessTicketCreateView`` (#2026)
-    - ``IrodsDataRequestModifyMixin.accept_request()`` always sets OK status for timeline event (#2027)
+    - ``IrodsDataRequestModifyMixin.accept_request()`` always sets OK status for timeline event (#2027, #2060)
+    - Accepting previously rejected iRODS data requests allowed (#2058)
 - **Taskflowbackend**
     - ``BatchValidateChecksumsTask`` file opening handling (#2049)